+++ Tests/Unit/Microsoft365DSC/Microsoft365DSC.EXOSharedMailbox.Tests.ps1
@@ -95,6 +95,10 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
             It 'Should return True from the Test method' {
                 Test-TargetResource @testParams | Should -Be $True
             }
+            It 'Should return false from the Test method' {
+                $testParams.PrimarySMTPAddress = 'test@contoso1.onmicrosoft.com'
+                Test-TargetResource @testParams | Should -Be $False
+            }
         }
 
         Context -Name 'Alias is Contained in the PrimarySMTP Address' -Fixture {
@@ -150,6 +154,33 @@ Describe -Name $Global:DscHelper.DescribeHeader -Fixture {
                     DisplayName        = 'Test Shared Mailbox'
                     PrimarySMTPAddress = 'Test@contoso.onmicrosoft.com'
                     EmailAddresses     = @('User1@contoso.onmicrosoft.com')
+                    Ensure             = 'Present'
+                    Credential         = $Credential
+                }
+
+                Mock -CommandName Get-Mailbox -MockWith {
+                    return @{
+                        Identity             = 'Test Shared Mailbox'
+                        RecipientTypeDetails = 'SharedMailbox'
+                        Alias                = 'test'
+                        EmailAddresses       = @('smtp:user@contoso.onmicrosoft.com', 'SMTP:test@contoso.onmicrosoft.com')
+                        PrimarySMTPAddress   = 'test@contoso.onmicrosoft.com'
+                    }
+                }
+            }
+
+            It 'Should call the Set method' {
+                Set-TargetResource @testParams
+            }
+        }
+
+        Context -Name 'Primary Smtp Address different' -Fixture {
+            BeforeAll {
+                $testParams = @{
+                    DisplayName        = 'Test Shared Mailbox'
+                    PrimarySMTPAddress = 'Test@contoso1.onmicrosoft.com'
+                    Alias                = 'test'
+                    EmailAddresses     = @('smtp:user@contoso.onmicrosoft.com', 'SMTP:test@contoso.onmicrosoft.com')
                     Ensure             = 'Present'
                     Credential         = $Credential
                 }