+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # UNRELEASED
 
+* EXOSharedMailbox
+  * Added capability to change the PrimarySMTPAddress of a Shared Mailbox
 * MISC
   * Updated AAD, EXO and Teams settings file to describe required roles.
   * Added a new personas documentation page to describe the targeted personas for the project.